Bug#37784391 - Fix failing mtr tests when running                with --hypergraph option [Windows]
Fix Windows failures

Failures are similar to that of Mac's. Carefullying examining
the failing diff shows that there isn't an actual failure.
But it is only because of the difference in the way "diff"
command treats the line endings that mtr reports it as
a failure. Using "diff" from "git bash" makes the failures
pass. So we give out a message to user when failing to
use "git bash" to get the right report.

Change-Id: I7f6e0a5d64688f63548abacb96c983a2e8755a3a